# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mena Wallet API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginRequest {
  pin: string;
}

export type LoginResponseRole =
  (typeof LoginResponseRole)[keyof typeof LoginResponseRole];

export const LoginResponseRole = {
  sender: "sender",
  receiver: "receiver",
} as const;

export interface LoginResponse {
  userId: string;
  name: string;
  role: LoginResponseRole;
  token: string;
}

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  received: "received",
} as const;

export interface Transaction {
  id: string;
  senderName: string;
  recipientName: string;
  amount: number;
  paidAmount: number;
  status: TransactionStatus;
  confirmedBy?: string | null;
  confirmedByName?: string | null;
  confirmedAt?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTransactionRequest {
  senderName: string;
  amount: number;
  paidAmount: number;
  notes?: string;
  createdByUserId: string;
}

export interface ConfirmTransactionRequest {
  confirmedByUserId: string;
  confirmedByName: string;
}

export interface Notification {
  id: string;
  userId: string;
  transactionId: string;
  title: string;
  body: string;
  isRead: boolean;
  createdAt: string;
}

export type FinancialReportTransactionsByPeriodItem = {
  period: string;
  amount: number;
  count: number;
};

export interface FinancialReport {
  totalAmount: number;
  totalPaid: number;
  totalPending: number;
  totalReceived: number;
  pendingCount: number;
  receivedCount: number;
  transactionsByPeriod: FinancialReportTransactionsByPeriodItem[];
}

export type GetTransactionsParams = {
  userId?: string;
  status?: GetTransactionsStatus;
  search?: string;
  minAmount?: number;
  maxAmount?: number;
  startDate?: string;
  endDate?: string;
};

export type GetTransactionsStatus =
  (typeof GetTransactionsStatus)[keyof typeof GetTransactionsStatus];

export const GetTransactionsStatus = {
  pending: "pending",
  received: "received",
  all: "all",
} as const;

export type GetNotificationsParams = {
  userId: string;
  unreadOnly?: boolean;
};

export type MarkAllNotificationsReadBody = {
  userId: string;
};

export type MarkAllNotificationsRead200 = {
  success: boolean;
};

export type GetFinancialReportParams = {
  userId?: string;
  period?: GetFinancialReportPeriod;
};

export type GetFinancialReportPeriod =
  (typeof GetFinancialReportPeriod)[keyof typeof GetFinancialReportPeriod];

export const GetFinancialReportPeriod = {
  week: "week",
  month: "month",
  year: "year",
  all: "all",
} as const;