# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mena Wallet API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with PIN
 */
export const LoginBody = zod.object({
  pin: zod.string(),
});

export const LoginResponse = zod.object({
  userId: zod.string(),
  name: zod.string(),
  role: zod.enum(["sender", "receiver"]),
  token: zod.string(),
});

/**
 * @summary Get all transactions
 */
export const GetTransactionsQueryParams = zod.object({
  userId: zod.coerce.string().optional(),
  status: zod.enum(["pending", "received", "all"]).optional(),
  search: zod.coerce.string().optional(),
  minAmount: zod.coerce.number().optional(),
  maxAmount: zod.coerce.number().optional(),
  startDate: zod.coerce.string().optional(),
  endDate: zod.coerce.string().optional(),
});

export const GetTransactionsResponseItem = zod.object({
  id: zod.string(),
  senderName: zod.string(),
  recipientName: zod.string(),
  amount: zod.number(),
  paidAmount: zod.number(),
  status: zod.enum(["pending", "received"]),
  confirmedBy: zod.string().nullish(),
  confirmedByName: zod.string().nullish(),
  confirmedAt: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const GetTransactionsResponse = zod.array(GetTransactionsResponseItem);

/**
 * @summary Create a new transaction
 */
export const CreateTransactionBody = zod.object({
  senderName: zod.string(),
  amount: zod.number(),
  paidAmount: zod.number(),
  notes: zod.string().optional(),
  createdByUserId: zod.string(),
});

/**
 * @summary Get transaction by ID
 */
export const GetTransactionParams = zod.object({
  id: zod.coerce.string(),
});

export const GetTransactionResponse = zod.object({
  id: zod.string(),
  senderName: zod.string(),
  recipientName: zod.string(),
  amount: zod.number(),
  paidAmount: zod.number(),
  status: zod.enum(["pending", "received"]),
  confirmedBy: zod.string().nullish(),
  confirmedByName: zod.string().nullish(),
  confirmedAt: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Confirm receipt of a transaction
 */
export const ConfirmTransactionParams = zod.object({
  id: zod.coerce.string(),
});

export const ConfirmTransactionBody = zod.object({
  confirmedByUserId: zod.string(),
  confirmedByName: zod.string(),
});

export const ConfirmTransactionResponse = zod.object({
  id: zod.string(),
  senderName: zod.string(),
  recipientName: zod.string(),
  amount: zod.number(),
  paidAmount: zod.number(),
  status: zod.enum(["pending", "received"]),
  confirmedBy: zod.string().nullish(),
  confirmedByName: zod.string().nullish(),
  confirmedAt: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get notifications for a user
 */
export const GetNotificationsQueryParams = zod.object({
  userId: zod.coerce.string(),
  unreadOnly: zod.coerce.boolean().optional(),
});

export const GetNotificationsResponseItem = zod.object({
  id: zod.string(),
  userId: zod.string(),
  transactionId: zod.string(),
  title: zod.string(),
  body: zod.string(),
  isRead: zod.boolean(),
  createdAt: zod.string(),
});
export const GetNotificationsResponse = zod.array(GetNotificationsResponseItem);

/**
 * @summary Mark notification as read
 */
export const MarkNotificationReadParams = zod.object({
  id: zod.coerce.string(),
});

export const MarkNotificationReadResponse = zod.object({
  id: zod.string(),
  userId: zod.string(),
  transactionId: zod.string(),
  title: zod.string(),
  body: zod.string(),
  isRead: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Mark all notifications as read for a user
 */
export const MarkAllNotificationsReadBody = zod.object({
  userId: zod.string(),
});

export const MarkAllNotificationsReadResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get financial report
 */
export const GetFinancialReportQueryParams = zod.object({
  userId: zod.coerce.string().optional(),
  period: zod.enum(["week", "month", "year", "all"]).optional(),
});

export const GetFinancialReportResponse = zod.object({
  totalAmount: zod.number(),
  totalPaid: zod.number(),
  totalPending: zod.number(),
  totalReceived: zod.number(),
  pendingCount: zod.number(),
  receivedCount: zod.number(),
  transactionsByPeriod: zod.array(
    zod.object({
      period: zod.string(),
      amount: zod.number(),
      count: zod.number(),
    }),
  ),
});