# lib/api-zod/src/generated/types/financialReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mena Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { FinancialReportTransactionsByPeriodItem } from "./financialReportTransactionsByPeriodItem";

export interface FinancialReport {
  totalAmount: number;
  totalPaid: number;
  totalPending: number;
  totalReceived: number;
  pendingCount: number;
  receivedCount: number;
  transactionsByPeriod: FinancialReportTransactionsByPeriodItem[];
}